Move commands into separate submodule
