+++ lychee-bin/src/commands/mod.rs
@@ -1,0 +1,5 @@
+pub(crate) mod check;
+pub(crate) mod dump;
+
+pub(crate) use check::check;
+pub(crate) use dump::dump;
